chore: rename NavBar to SampleComponent
