+++ jest.config.js
@@ -8,7 +8,7 @@ module.exports = {
 
   setupFilesAfterEnv: ['./script/test/jest-setup.ts'],
   projects: [
-    '<rootDir>/packages/basics/locale/',
+    '<rootDir>/packages/basics/locale',
     '<rootDir>/packages/composites/ui-atoms',
     '<rootDir>/packages/plugins/components',
     '<rootDir>/packages/zapp/console',
