+++ packages/plugins/components/src/index.ts
@@ -1,2 +1,1 @@
-export { NavBar } from './NavBar';
 export { AppInfo, type VersionInfo } from './AppInfo';
